+++ assets/composition/walking-skeleton/wc-design-system/templates/design-system/src/modules.d.ts
@@ -1,0 +1,3 @@
+declare module '*.css';
+declare module '@rgoussu.dev/planks/tokens';
+declare module '@rgoussu.dev/planks/styles';